Fix supporting 'iceberg.catalog' table property in Hive QL.
